Reorganized code for Planner to separate frequently modified code from base classes
